requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.1.6
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
docopt==0.6.2
idna==3.4
pipreqs==0.4.11
PyJWT==1.7.1
pytz==2022.7.1
requests==2.28.2
sqlparse==0.4.3
urllib3==1.26.14
yarg==0.1.9
